Fixed login bug by creating login redirect #90
